requirements.txt
langchain-google-genai
langgraph
python-dotenv
google-generativeai
langchain-core